PathExplorers CRUD refactoring
Moved CRUD functions in extensions
