+++ Sources/Scout/Implementations/Serialization/PathExplorerSerialization+Get.swift
@@ -1,0 +1,89 @@
+extension PathExplorerSerialization {
+
+    func get(for key: String) throws -> Self {
+        guard let dict = value as? DictionaryValue else {
+            throw PathExplorerError.dictionarySubscript(readingPath)
+        }
+
+        guard let childValue = dict[key] else {
+            throw PathExplorerError.subscriptMissingKey(path: readingPath, key: key, bestMatch: key.bestJaroWinklerMatchIn(propositions: Set(dict.keys)))
+        }
+
+        return PathExplorerSerialization(value: childValue, path: readingPath.appending(key))
+    }
+
+    /// - parameter negativeIndexEnabled: If set to `true`, it is possible to get the last element of an array with the index `-1`
+    func get(at index: Int, negativeIndexEnabled: Bool = true) throws -> Self {
+        guard let array = value as? ArrayValue else {
+            throw PathExplorerError.arraySubscript(readingPath)
+        }
+
+        if index == -1, negativeIndexEnabled {
+            if array.isEmpty {
+                throw PathExplorerError.subscriptWrongIndex(path: readingPath, index: index, arrayCount: array.count)
+            }
+            return PathExplorerSerialization(value: array[array.count - 1], path: readingPath.appending(index))
+        }
+
+        guard array.count > index, index >= 0 else {
+            throw PathExplorerError.subscriptWrongIndex(path: readingPath, index: index, arrayCount: array.count)
+        }
+
+        return PathExplorerSerialization(value: array[index], path: readingPath.appending(index))
+    }
+
+    /// - Returns: The count of the array if  `value` is an array
+    func getChildrenCount() throws -> Self {
+        if let arrayValue = value as? ArrayValue {
+            return PathExplorerSerialization(value: arrayValue.count, path: readingPath.appending(.count))
+        } else if let dictValue = value as? DictionaryValue {
+            return PathExplorerSerialization(value: dictValue.count, path: readingPath.appending(.count))
+        } else {
+            throw PathExplorerError.arrayCountWrongUsage(path: readingPath)
+        }
+    }
+
+    /// - parameter negativeIndexEnabled: If set to `true`, it is possible to get the last element of an array with the index `-1`
+    func get(element: PathElement, negativeIndexEnabled: Bool = true) throws -> Self {
+        guard readingPath.last != .count else {
+            throw PathExplorerError.arrayCountWrongUsage(path: readingPath)
+        }
+
+        switch element {
+        case .key(let key): return try get(for: key)
+        case .index(let index): return try get(at: index, negativeIndexEnabled: negativeIndexEnabled)
+        case .count: return try getChildrenCount()
+        }
+    }
+
+    public func get(_ path: Path) throws -> Self {
+        var currentPathExplorer = self
+
+        try path.forEach { element in
+            currentPathExplorer = try currentPathExplorer.get(element: element)
+        }
+
+        return currentPathExplorer
+    }
+
+    /// Explorer the path in parameter to find each `PathExplorer` in
+    /// - Parameter path: The path to explore. Should not be empty
+    /// - Throws: If the path is invalid
+    /// - Returns: The explorers discovered, the path without the last element, and the last element
+    func getExplorers(from path: Path) throws -> (explorers: [Self], path: Path, lastElement: PathElement) {
+        assert(!path.isEmpty)
+
+        var craftingPath = path
+        let lastElement = craftingPath.removeLast()
+
+        let explorers = try craftingPath.reduce([self]) { (explorers, element) in
+            guard let currentExplorer = try explorers.last?.get(element: element) else {
+                return explorers // should not happen
+            }
+
+            return explorers + [currentExplorer]
+        }
+
+        return (explorers, craftingPath, lastElement)
+    }
+}
